Fixed detected race conditions in udp receiver package
